# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Server API for Hamster Money Service Application",
        "title": "Hamster API",
        "contact": {
            "name": "Hamster API Support",
            "url": "https://t.me/CodeMaster482",
            "email": "grigorikovalenko@gmail.com"
        },
        "version": "1.0.1"
    },
    "host": "localhost:8080",
    "basePath": "/user/{userID}/account/feed",
    "paths": {
        "/api/account/create": {
            "post": {
                "description": "Create account",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Create account",
                "parameters": [
                    {
                        "description": "Input account create",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.CreateAccount"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Create account",
                        "schema": {
                            "$ref": "#/definitions/http.Response-http_AccountCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Client error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "403": {
                        "description": "Forbidden user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/account/update": {
            "put": {
                "description": "Put \taccount",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "PUT \tUpdate",
                "parameters": [
                    {
                        "description": "Input transactin update",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.UpdateAccount"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update account",
                        "schema": {
                            "$ref": "#/definitions/http.Response-http_NilBody"
                        }
                    },
                    "400": {
                        "description": "Client error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "403": {
                        "description": "Forbidden user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/account/{account_id}/delete": {
            "delete": {
                "description": "Delete account with chosen ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Delete Account",
                "responses": {
                    "200": {
                        "description": "Account deleted",
                        "schema": {
                            "$ref": "#/definitions/http.Response-http_NilBody"
                        }
                    },
                    "400": {
                        "description": "Account error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "401": {
                        "description": "User unathorized",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "403": {
                        "description": "User hasn't rights",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "http.AccountCreateResponse": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                }
            }
        },
        "http.CreateAccount": {
            "type": "object",
            "properties": {
                "accumulation": {
                    "type": "boolean"
                },
                "balance": {
                    "type": "number"
                },
                "balance_enabled": {
                    "type": "boolean"
                },
                "mean_payment": {
                    "type": "string"
                }
            }
        },
...
        "http.NilBody": {
            "type": "object"
        },
...
        "http.Response-http_AccountCreateResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "$ref": "#/definitions/http.AccountCreateResponse"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "http.Response-http_NilBody": {
            "type": "object",
            "properties": {
                "body": {
                    "$ref": "#/definitions/http.NilBody"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "http.ResponseError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "http.UpdateAccount": {
            "type": "object",
            "properties": {
                "accumulation": {
                    "type": "boolean"
                },
                "balance": {
                    "type": "number"
                },
                "balance_enabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "mean_payment": {
                    "type": "string"
                }
            }
        },
...
    }
}